backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
alembic==1.13.0
psycopg2-binary==2.9.9
greenlet==3.2.3
# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Testing
httpx==0.25.2
